Refactor
  - move each resources implementation to specific files
